+++ java/src/main/java/ifeed/problem/assigning/filtersWithException/EmptyOrbitWithException.java
@@ -32,14 +32,14 @@ public class EmptyOrbitWithException extends EmptyOrbit {
     }
 
     public void setOrbitException(Set<Integer> exception){
-        if(exception.size() >= super.params.getRightSetCardinality() / 2){
+        if(exception.size() >= (double) super.params.getRightSetCardinality() / 2){
             throw new IllegalStateException("The number of exceptions should be smaller than half the number of the valid instances");
         }
         this.orbitException = exception;
     }
 
     public void setInstrumentException(Set<Integer> exception){
-        if(exception.size() >= super.params.getLeftSetCardinality() / 2){
+        if(exception.size() >= (double) super.params.getLeftSetCardinality() / 2){
             throw new IllegalStateException("The number of exceptions should be smaller than half the number of the valid instances");
         }
         this.instrumentException = exception;
