+++ java/src/main/java/ifeed/problem/assigning/filtersWithException/NotInOrbitWithException.java
@@ -57,7 +57,7 @@ public class NotInOrbitWithException extends NotInOrbit {
             }
         }
 
-        if(exception.size() >= super.params.getRightSetCardinality() / 2){
+        if(exception.size() >= (double) super.params.getRightSetCardinality() / 2){
             throw new IllegalStateException("The number of exceptions should be smaller than half the number of the valid instances");
         }
         this.orbitException = exception;
